Add files via upload
This is Project on Excel Sheet where a data set of particular store named Vrindha Store is taken and has the data about the orders and sales of products in the year 2022. In this I have converted the data into pictorial representation like pie chart, bar graph etc. The major change in analysis is done with the help of pivot tables and pivot charts.